Shared EventRow: /speaking uses the /events row style with flags
Extract the event row from EventsFilter into components/events/EventRow.tsx
and the flag lookup into lib/flags.ts. /speaking's 'Next up' strip now
renders the exact same component server-side (no hydration), so both pages
stay visually identical — date block, type/featured/video badges, country
flags, chevron, hover glow. Type chips are tokenized (workshop→signal,
meetup→warn, default→accent).
